remove ut.isep.management.repository.pgsql package, extract contents to parent directory
