+++ management-server/src/main/kotlin/ut/isep/management/DummyDataLoader.kt
@@ -3,8 +3,8 @@ package ut.isep.management
 import org.springframework.boot.CommandLineRunner
 import org.springframework.stereotype.Component
 import ut.isep.management.model.entity.*
-import ut.isep.management.repository.pgsql.AssignmentRepository
-import ut.isep.management.repository.pgsql.SectionRepository
+import ut.isep.management.repository.AssignmentRepository
+import ut.isep.management.repository.SectionRepository
 
 @Component
 class DummyDataLoader(
