+++ management-server/src/main/kotlin/ut/isep/management/repository/AssignmentRepository.kt
@@ -1,4 +1,4 @@
-package ut.isep.management.repository.pgsql
+package ut.isep.management.repository
 
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
