+++ management-server/src/main/kotlin/ut/isep/management/repository/PostgreSQLConfig.kt
@@ -1,4 +1,4 @@
-package ut.isep.management.repository.pgsql
+package ut.isep.management.repository
 
 import org.springframework.boot.autoconfigure.jdbc.DataSourceProperties
 import org.springframework.boot.context.properties.ConfigurationProperties
