+++ management-server/src/main/kotlin/ut/isep/management/service/SectionService.kt
@@ -4,7 +4,7 @@ import dto.*
 import jakarta.transaction.Transactional
 import org.springframework.stereotype.Service
 import ut.isep.management.model.entity.*
-import ut.isep.management.repository.pgsql.SectionRepository
+import ut.isep.management.repository.SectionRepository
 import java.util.*
